Fix suffix string generation
